+++ modules/account-lib/src/index.ts
@@ -146,6 +146,9 @@ export { Coredao };
 import * as Apt from '@bitgo/sdk-coin-apt';
 export { Apt };
 
+import * as Icp from '@bitgo/sdk-coin-icp';
+export { Icp };
+
 const coinBuilderMap = {
   trx: Trx.WrappedBuilder,
   ttrx: Trx.WrappedBuilder,
@@ -222,6 +225,8 @@ const coinBuilderMap = {
   tcoredao: Coredao.TransactionBuilder,
   apt: Apt.TransactionBuilder,
   tapt: Apt.TransactionBuilder,
+  icp: Icp.TransactionBuilder,
+  ticp: Icp.TransactionBuilder,
   baby: Baby.TransactionBuilder,
   tbaby: Baby.TransactionBuilder,
 };