+++ modules/statics/src/networks.ts
@@ -1149,7 +1149,70 @@ class CoredaoTestnet extends Testnet implements EthereumNetwork {
   forwarderFactoryAddress = '0x37996e762fa8b671869740c79eb33f625b3bf92a';
   forwarderImplementationAddress = '0xd5fe1c1f216b775dfd30638fa7164d41321ef79b';
 }
-
+class Xdc extends Mainnet implements EthereumNetwork {
+  name = 'XdcChain';
+  family = CoinFamily.XDC;
+  explorerUrl = 'https://xdcscan.com/tx';
+  accountExplorerUrl = 'https://xdcscan.com/address';
+  chainId = 50;
+  nativeCoinOperationHashPrefix = '50';
+}
+class XdcTestnet extends Testnet implements EthereumNetwork {
+  name = 'XdcChainTestnet';
+  family = CoinFamily.XDC;
+  explorerUrl = 'https://testnet.xdcscan.com/tx';
+  accountExplorerUrl = 'https://testnet.xdcscan.com/address';
+  chainId = 51;
+  nativeCoinOperationHashPrefix = '51';
+}
+class Wemix extends Mainnet implements EthereumNetwork {
+  name = 'WemixChain';
+  family = CoinFamily.WEMIX;
+  explorerUrl = 'https://explorer.wemix.com/tx';
+  accountExplorerUrl = 'https://explorer.wemix.com/address';
+  chainId = 1111;
+  nativeCoinOperationHashPrefix = '1111';
+}
+class WemixTestnet extends Testnet implements EthereumNetwork {
+  name = 'WemixChainTestnet';
+  family = CoinFamily.WEMIX;
+  explorerUrl = 'https://explorer.test.wemix.com/tx';
+  accountExplorerUrl = 'https://explorer.test.wemix.com/address';
+  chainId = 1112;
+  nativeCoinOperationHashPrefix = '1112';
+}
+class Flare extends Mainnet implements EthereumNetwork {
+  name = 'Flarechain';
+  family = CoinFamily.FLR;
+  explorerUrl = 'https://flare-explorer.flare.network/tx';
+  accountExplorerUrl = 'https://flare-explorer.flare.network/address';
+  chainId = 14;
+  nativeCoinOperationHashPrefix = '14';
+}
+class FlareTestnet extends Testnet implements EthereumNetwork {
+  name = 'FlarechainTestnet';
+  family = CoinFamily.FLR;
+  explorerUrl = 'https://coston2-explorer.flare.network/tx';
+  accountExplorerUrl = 'https://coston2-explorer.flare.network/address';
+  chainId = 114;
+  nativeCoinOperationHashPrefix = '114';
+}
+class Songbird extends Mainnet implements EthereumNetwork {
+  name = 'Songbirdchain';
+  family = CoinFamily.SGB;
+  explorerUrl = 'https://songbird-explorer.flare.network/tx';
+  accountExplorerUrl = 'https://songbird-explorer.flare.network/address';
+  chainId = 19;
+  nativeCoinOperationHashPrefix = '19';
+}
+class SongbirdTestnet extends Testnet implements EthereumNetwork {
+  name = 'SongbirdchainTestnet';
+  family = CoinFamily.SGB;
+  explorerUrl = 'https://coston-explorer.flare.network/tx';
+  accountExplorerUrl = 'https://coston-explorer.flare.network/address';
+  chainId = 16;
+  nativeCoinOperationHashPrefix = '16';
+}
 class BaseChainTestnet extends Testnet implements EthereumNetwork {
   name = 'BaseChainTestnet';
   family = CoinFamily.BASEETH;
@@ -1210,6 +1273,7 @@ export const Networks = {
     ethereumClassic: Object.freeze(new EthereumClassic()),
     ethereumW: Object.freeze(new EthereumW()),
     fiat: Object.freeze(new Fiat()),
+    flr: Object.freeze(new Flare()),
     hash: Object.freeze(new Hash()),
     hedera: Object.freeze(new Hedera()),
     icp: Object.freeze(new Icp()),
@@ -1227,6 +1291,7 @@ export const Networks = {
     rune: Object.freeze(new Rune()),
     stellar: Object.freeze(new Stellar()),
     sei: Object.freeze(new Sei()),
+    sgb: Object.freeze(new Songbird()),
     sol: Object.freeze(new Sol()),
     sui: Object.freeze(new Sui()),
     near: Object.freeze(new Near()),
@@ -1236,6 +1301,8 @@ export const Networks = {
     tia: Object.freeze(new Tia()),
     ton: Object.freeze(new Ton()),
     trx: Object.freeze(new Trx()),
+    wemix: Object.freeze(new Wemix()),
+    xdc: Object.freeze(new Xdc()),
     xrp: Object.freeze(new Xrp()),
     xtz: Object.freeze(new Xtz()),
     zCash: Object.freeze(new ZCash()),
@@ -1271,6 +1338,7 @@ export const Networks = {
     eCash: Object.freeze(new ECashTestnet()),
     eos: Object.freeze(new EosTestnet()),
     fiat: Object.freeze(new FiatTestnet()),
+    flr: Object.freeze(new FlareTestnet()),
     pyrmont: Object.freeze(new Pyrmont()),
     ethereumClassicTestnet: Object.freeze(new EthereumClassicTestnet()),
     hash: Object.freeze(new HashTestnet()),
@@ -1291,6 +1359,7 @@ export const Networks = {
     osmo: Object.freeze(new OsmoTestnet()),
     rbtc: Object.freeze(new RbtcTestnet()),
     rune: Object.freeze(new RuneTestNet()),
+    sgb: Object.freeze(new SongbirdTestnet()),
     stellar: Object.freeze(new StellarTestnet()),
     sei: Object.freeze(new SeiTestnet()),
     sol: Object.freeze(new SolTestnet()),
@@ -1303,6 +1372,8 @@ export const Networks = {
     tia: Object.freeze(new TiaTestnet()),
     ton: Object.freeze(new TonTestnet()),
     trx: Object.freeze(new TrxTestnet()),
+    wemix: Object.freeze(new WemixTestnet()),
+    xdc: Object.freeze(new XdcTestnet()),
     xrp: Object.freeze(new XrpTestnet()),
     xtz: Object.freeze(new XtzTestnet()),
     zCash: Object.freeze(new ZCashTestnet()),