+++ modules/statics/test/unit/coins.ts
@@ -261,6 +261,20 @@ const custodyFeatures: Record<string, { features: CoinFeature[] }> = {
   'sol:wen': { features: [CoinFeature.CUSTODY_BITGO_FRANKFURT] },
   'sol:nos': { features: [CoinFeature.CUSTODY_BITGO_FRANKFURT] },
   'sol:spx': { features: [CoinFeature.CUSTODY_BITGO_FRANKFURT] },
+  'sol:trump': {
+    features: [
+      CoinFeature.CUSTODY_BITGO_FRANKFURT,
+      CoinFeature.CUSTODY_BITGO_GERMANY,
+      CoinFeature.CUSTODY_BITGO_SINGAPORE,
+    ],
+  },
+  'sol:melania': {
+    features: [
+      CoinFeature.CUSTODY_BITGO_FRANKFURT,
+      CoinFeature.CUSTODY_BITGO_GERMANY,
+      CoinFeature.CUSTODY_BITGO_SINGAPORE,
+    ],
+  },
   'xlm:ZUSD-GDF6VOEGRWLOZ64PQQGKD2IYWA22RLT37GJKS2EJXZHT2VLAGWLC5TOB': {
     features: [CoinFeature.CUSTODY_BITGO_FRANKFURT],
   },