+++ modules/abstract-utxo/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [9.12.0](https://github.com/BitGo/BitGoJS/compare/@bitgo/abstract-utxo@9.11.1...@bitgo/abstract-utxo@9.12.0) (2025-01-20)
+
+### Bug Fixes
+
+- **abstract-utxo:** improve createHalfSigned ([bac55b4](https://github.com/BitGo/BitGoJS/commit/bac55b4aa5a9c6624479e55424568a2bda32cac9))
+
+### Features
+
+- **abstract-utxo:** add descriptor support for offline vault ([7082893](https://github.com/BitGo/BitGoJS/commit/7082893036763d6c5bbe37236ebac68acd1976f5))
+
 ## [9.11.1](https://github.com/BitGo/BitGoJS/compare/@bitgo/abstract-utxo@9.11.0...@bitgo/abstract-utxo@9.11.1) (2025-01-15)
 
 **Note:** Version bump only for package @bitgo/abstract-utxo