+++ modules/sdk-coin-apt/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.1](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-apt@1.4.0...@bitgo/sdk-coin-apt@1.4.1) (2025-01-20)
+
+**Note:** Version bump only for package @bitgo/sdk-coin-apt
+
 # [1.4.0](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-apt@1.3.9...@bitgo/sdk-coin-apt@1.4.0) (2025-01-15)
 
 ### Features