+++ modules/sdk-coin-ltc/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.1.10](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-ltc@3.1.9...@bitgo/sdk-coin-ltc@3.1.10) (2025-01-20)
+
+**Note:** Version bump only for package @bitgo/sdk-coin-ltc
+
 ## [3.1.9](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-ltc@3.1.8...@bitgo/sdk-coin-ltc@3.1.9) (2025-01-15)
 
 **Note:** Version bump only for package @bitgo/sdk-coin-ltc