+++ modules/sdk-coin-xtz/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.58](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-xtz@2.1.57...@bitgo/sdk-coin-xtz@2.1.58) (2025-01-20)
+
+**Note:** Version bump only for package @bitgo/sdk-coin-xtz
+
 ## [2.1.57](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-xtz@2.1.56...@bitgo/sdk-coin-xtz@2.1.57) (2025-01-15)
 
 **Note:** Version bump only for package @bitgo/sdk-coin-xtz